tests fixes
