+++ src/components/AccordionTableRow/AccordionTableRow.test.tsx
@@ -4,7 +4,7 @@ import { AccordionTableRow } from './';
 
 describe('<AccordionTableRow />', () => {
   it('should render an empty <AccordionTableRow> component', () => {
-    const wrapper = enzyme.shallow(<AccordionTableRow />);
-    expect(wrapper.text()).toBe('<styled.tr />');
+    const wrapper = enzyme.mount(<AccordionTableRow />);
+    expect(wrapper.find('tr').length > 0).toBeTruthy();
   });
 });
