+++ src/components/BodyText/BodyText.test.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import { BodyText } from './';
 import 'jest-styled-components';
 
 describe('<BodyText />', () => {
   it('should render default <BodyText> component', () => {
-    const wrapper = shallow(
+    const wrapper = mount(
       <BodyText>
         Dolore id excepteur id aute amet enim aute esse eu quis anim aliquip officia. Do Lorem sit quis anim velit
         cillum magna aliquip aliquip officia cupidatat ex. Cupidatat esse labore cupidatat est non adipisicing
@@ -16,7 +16,7 @@ describe('<BodyText />', () => {
   });
 
   it('text should have right alignment', () => {
-    const wrapper = shallow(
+    const wrapper = mount(
       <BodyText align="right">
         Dolore id excepteur id aute amet enim aute esse eu quis anim aliquip officia. Do Lorem sit quis anim velit
         cillum magna aliquip aliquip officia cupidatat ex. Cupidatat esse labore cupidatat est non adipisicing
