+++ src/components/Sticky/sticky.test.tsx
@@ -5,11 +5,11 @@ import { Sticky } from './';
 
 describe('<Sticky />', () => {
   it('should render sticky sticked to the bottom by default', () => {
-    const wrapper = enzyme.shallow(<Sticky>Child</Sticky>);
+    const wrapper = enzyme.mount(<Sticky>Child</Sticky>);
     expect(wrapper).toHaveStyleRule('bottom', '0');
   });
   it('should render sticky sticked to the left', () => {
-    const wrapper = enzyme.shallow(<Sticky position="left">Child</Sticky>);
+    const wrapper = enzyme.mount(<Sticky position="left">Child</Sticky>);
     expect(wrapper).toHaveStyleRule('left', '0');
   });
 });
